# openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ErrorCode": {
        "description": "Stable, machine-readable error identifiers.\n\nThese are part of the public API contract: clients branch on them. Add new\nmembers freely; never rename or repurpose an existing one.",
        "enum": [
          "internal_error",
          "not_found",
          "conflict",
          "validation_failed",
          "unauthenticated",
          "forbidden",
          "rate_limited",
          "upstream_failed",
          "service_unavailable",
          "configuration_error"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Stable machine-readable identifier."
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "description": "Human-readable explanation. Not for parsing.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Correlates this response with the server logs.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FilterDecisionRead": {
        "description": "One verdict from the filtering agent.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_fallback": {
            "description": "True when the agent failed and this is the fail-closed default, not a real judgement.",
            "title": "Is Fallback",
            "type": "boolean"
          },
          "keep": {
            "title": "Keep",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "id",
          "keep",
          "category",
          "is_fallback"
        ],
        "title": "FilterDecisionRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestionConfigResponse": {
        "description": "The knobs the ingestion pipeline is currently running with.",
        "properties": {
          "embedding_dim": {
            "title": "Embedding Dim",
            "type": "integer"
          },
          "embedding_executor": {
            "title": "Embedding Executor",
            "type": "string"
          },
          "embedding_model": {
            "title": "Embedding Model",
            "type": "string"
          },
          "embedding_workers": {
            "title": "Embedding Workers",
            "type": "integer"
          },
          "filter_system_prompt": {
            "title": "Filter System Prompt",
            "type": "string"
          },
          "llm_provider": {
            "title": "Llm Provider",
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "platform",
          "filter_system_prompt",
          "prompt_version",
          "llm_provider",
          "embedding_model",
          "embedding_dim",
          "embedding_executor",
          "embedding_workers",
          "source"
        ],
        "title": "IngestionConfigResponse",
        "type": "object"
      },
      "IngestionRunRequest": {
        "description": "Optional per-run overrides for the scheduler.",
        "properties": {
          "dry_run": {
            "default": false,
            "description": "Run source, filtering and embedding, then roll back instead of persisting.",
            "title": "Dry Run",
            "type": "boolean"
          },
          "limit": {
            "anyOf": [
              {
                "maximum": 1000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cap on messages pulled this run.",
            "title": "Limit"
          },
          "system_prompt_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Overrides INGESTION_FILTER_SYSTEM_PROMPT for this run only. Intended for tuning against dry_run; changing retention policy for a persisted run should be a configuration change, not a request parameter.",
            "title": "System Prompt Override"
          }
        },
        "title": "IngestionRunRequest",
        "type": "object"
      },
...
      "IngestionRunSummary": {
        "description": "One past run, as the history list shows it.\n\nBuilt from the stored row rather than from the in-memory result, so the\nhistory screen and a run's own response cannot drift apart.",
        "properties": {
          "already_ingested": {
            "title": "Already Ingested",
            "type": "integer"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/FilterDecisionRead"
            },
            "title": "Decisions",
            "type": "array"
          },
          "discarded": {
            "title": "Discarded",
            "type": "integer"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "embedded": {
            "title": "Embedded",
            "type": "integer"
          },
          "embedding_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "evaluated": {
            "title": "Evaluated",
            "type": "integer"
          },
          "fetched": {
            "title": "Fetched",
            "type": "integer"
          },
          "filter_errors": {
            "title": "Filter Errors",
            "type": "integer"
          },
          "filter_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filter Provider"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "persisted": {
            "title": "Persisted",
            "type": "integer"
          },
          "platform": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Platform"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null only for runs recorded before per-platform pipelines."
          },
          "retained": {
            "title": "Retained",
            "type": "integer"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "description": "\"success\", \"partial\" or \"failed\".",
            "title": "Status",
            "type": "string"
          },
          "users_provisioned": {
            "title": "Users Provisioned",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "started_at",
          "duration_ms",
          "dry_run",
          "status",
          "fetched",
          "already_ingested",
          "evaluated",
          "retained",
          "discarded",
          "embedded",
          "persisted",
          "users_provisioned",
          "filter_errors"
        ],
        "title": "IngestionRunSummary",
        "type": "object"
      },
...
      "Platform": {
        "description": "Third-party systems we can ingest identities and messages from.",
        "enum": [
          "slack",
          "github",
          "teams",
          "email",
          "linear",
          "other"
        ],
        "title": "Platform",
        "type": "string"
      },
      "QueuedRunResponse": {
        "description": "What a trigger returns once ingestion is durably orchestrated.\n\nThe run has been *accepted*, not completed - with a local model a real run\ntakes minutes, so blocking the request until it finished would time out in\nthe browser long before there was anything to report. Poll\n`GET /ingestion/runs/{platform}/{run_id}` for progress.",
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "run_id": {
            "description": "Poll for this run's status with it.",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "description": "Always 'queued' here.",
            "title": "Status",
            "type": "string"
          },
          "workflow_id": {
            "description": "The durable workflow, visible in the Temporal UI.",
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "platform",
          "workflow_id"
        ],
        "title": "QueuedRunResponse",
        "type": "object"
      },
      "RunProgressResponse": {
        "description": "A run in flight, or its final counters once it has finished.\n\n`stage` is the pipeline step currently executing. Before this existed the\nconsole animated fake progress, because the API had nothing to report until\nthe whole run completed.",
        "properties": {
          "embedded": {
            "default": 0,
            "title": "Embedded",
            "type": "integer"
          },
          "evaluated": {
            "default": 0,
            "title": "Evaluated",
            "type": "integer"
          },
          "fetched": {
            "default": 0,
            "title": "Fetched",
            "type": "integer"
          },
          "filtered": {
            "default": 0,
            "title": "Filtered",
            "type": "integer"
          },
          "persisted": {
            "default": 0,
            "title": "Persisted",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngestionRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "stage": {
            "default": "starting",
            "description": "Pipeline step in flight.",
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "description": "queued, running, completed, failed or cancelled.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunProgressResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Service credential for machine callers (scheduler, backend jobs). The legacy `X-Cron-Token` header is also accepted. Grants whatever scopes were issued with the key, e.g. `ingest:run`.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "DevUser": {
        "description": "Local development only: impersonate a user by name. Defaults to `insights:read, messages:read`; add the `X-Dev-Scopes` header (comma-separated) to override. Disabled whenever DEV_AUTH_ENABLED=false.",
        "in": "header",
        "name": "X-Dev-User",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Versioned under `/api/{version}`. Probes (`/health`, `/ready`) are unversioned infrastructure contracts. Every error shares one envelope; see the `ErrorResponse` schema.",
    "title": "threadline API v1",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ingestion/config/{platform}": {
      "get": {
        "description": "Confirms which prompt, model, executor and connector a deployment picked up.",
        "operationId": "get_ingestion_config",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Platform"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required or invalid."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated but not permitted."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "DevUser": []
          }
        ],
        "summary": "Inspect one platform's active ingestion configuration",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/ingestion/runs": {
      "get": {
        "description": "Newest first, each with the filtering verdicts it recorded. That trail is what makes tuning the filter policy a reviewable act rather than a guess - it is the evidence behind every retention decision.\n\nDry runs appear here too. They deliberately store no messages, but the record of what they *would* have stored is the entire point of running one.\n\nUnfiltered by default, across every platform. Pass `platform` to see just one pipeline's history.",
        "operationId": "list_ingestion_runs",
        "parameters": [
          {
            "description": "How many runs to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "How many runs to return.",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Platform"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngestionRunSummary"
                  },
                  "title": "Response List Ingestion Runs",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required or invalid."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated but not permitted."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "DevUser": []
          }
        ],
        "summary": "List past ingestion runs",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/ingestion/runs/{platform}": {
      "post": {
        "description": "One pipeline per platform, invoked by the external scheduler independently per platform. Pulls messages from that platform's connector, filters them with the agentic policy, generates embeddings, and stores the survivors.\n\n**Accepted, not completed.** The run is handed to a durable workflow and this returns immediately with a `run_id`. Poll `GET /ingestion/runs/{platform}/{run_id}` for stage-by-stage progress and the final counters. Inference against a local model takes minutes, so waiting for it in the request would time out in a browser long before there was anything to report.\n\nIdempotent twice over: messages already stored are skipped by `(platform, external_message_id)`, and the workflow id derived from `run_id` means the same run cannot be started concurrently.\n\n404 if no connector is configured for the platform yet.",
        "operationId": "run_ingestion",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Platform"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/IngestionRunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuedRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required or invalid."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated but not permitted."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "DevUser": []
          }
        ],
        "summary": "Queue one platform's ingestion cycle",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/ingestion/runs/{platform}/{run_id}": {
      "get": {
        "description": "Stage-by-stage progress while a run is in flight, and the full report once it finishes. `status` is one of queued, running, completed, failed or cancelled; `result` is populated only on completion.\n\nProgress comes from a workflow query, which reads live state without replaying history, so polling this is cheap.",
        "operationId": "get_run_status",
        "parameters": [
          {
            "description": "Which platform's pipeline to run.",
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Platform",
              "description": "Which platform's pipeline to run."
            }
          },
          {
            "description": "The id returned when the run was queued.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "The id returned when the run was queued.",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required or invalid."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authenticated but not permitted."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "DevUser": []
          }
        ],
        "summary": "Poll a queued run",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  }
}